Refactor : section 관련 control 분리
<description>
section의 데이터를 다루는 함수들을 section-control.js파일로 옮겼다.
